(PC-14840)[API] chore: remove deprecated code for legacy cron
